Move AppConfig to :models
