+++ common/models/src/main/java/app/ss/models/config/AppConfig.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021. Adventech <info@adventech.io>
+ * Copyright (c) 2023. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package com.cryart.sabbathschool.core.model
+package app.ss.models.config
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
